--- packages/apps-config/src/ipfs-endpoints/types.ts
@@ -1,8 +1,0 @@
-// Copyright 2017-2021 @polkadot/apps-config authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-export interface AuthIpfsEndpoint {
-  text?: string;
-  value: string;
-  location?: string;
-}
