+++ packages/apps-config/src/ipfs-gateway-endpoints/types.ts
@@ -1,0 +1,8 @@
+// Copyright 2019-2021 Crust Network authors & contributors
+// License: Apache-2.0
+
+export interface AuthIpfsEndpoint {
+  text?: string;
+  value: string;
+  location?: string;
+}
